+++ lib/traces.rb
@@ -5,8 +5,10 @@
 
 require_relative "traces/version"
 require_relative "traces/provider"
-require_relative "traces/context"
 
 # @namespace
 module Traces
+	if self.enabled?
+		Config::DEFAULT.prepare
+	end
 end