+++ lib/traces/provider.rb
@@ -36,8 +36,6 @@ module Traces
 			
 			return provider
 		end
-		
-		Config::DEFAULT.prepare
 	else
 		def self.Provider(klass, &block)
 			# Tracing disabled.